🩹 (fetcher.js): skip log entry when save fails
This commit prevents the fetcher.js module to add a new row to the log
sheet even if the file was not successfully saved to Google Drive.

The fix only adding a new row to the log if the return value of the
`createViewFileFormula()` is not empty. This ensures that the log file
reflects the actual status of the file saving operation.

Fixes: #84