🐛 fix(ci): remove deprecated CodeQL Action workflow file
This commit removes the deprecated CodeQL Action workflow file as it is
no longer updated or supported. The deprecated version (v1) of the
CodeQL Action was officially deprecated on January 18th, 2023, and
upgrading to v2 is recommended for better performance, improved security,
and new features.

For more information about the deprecation and the benefits of upgrading,
please refer to the following link:
https://github.blog/changelog/2023-01-18-code-scanning-codeql-action-v1-is-now-deprecated/

Fixes: #98