🚑 fix(gh-page): Correctly run custom script for consent UI in production
This commit addresses an issue where the custom script for the consent UI was not running correctly in production environments due to the presence of the `data-ampdevmode` attribute.

Two new settings have been added to the site configuration according to the latest theme version 2.9.1:

- `consent_dev_mode`: This setting can be set to `false` to remove the `data-ampdevmode` attribute from the `amp-script` component, ensuring it runs correctly in production.

- `script_hash`: This setting holds the script hash for the `get_stored_consentStates` script, which is used for the consent UI. This hash is necessary for the script to run correctly.

The `data-ampdevmode` attribute has been removed from the `amp-script` component in non-development environments to ensure the script runs as expected.